+++ packages/x-charts-premium/src/CandlestickChart/CandlestickPlot.tsx
@@ -5,7 +5,7 @@ import { useXScale, useYScale } from '@mui/x-charts/hooks';
 import { type DefaultizedOHLCSeriesType } from '../models';
 import { useOHLCSeriesContext } from '../hooks/useOHLCSeries';
 import { useCandlestickPlotData } from './useCandlestickPlotData';
-import { useWebGLLayer } from '../ChartsWebGLLayer/ChartsWebGLLayer';
+import { useWebGLLayer } from '../ChartsWebGLLayer/ChartsWebGLContext';
 import { checkCandlestickScaleErrors } from './checkCandlestickScaleErrors';
 import { CandlestickWebGLProgram } from './CandlestickWebGLProgram';
