+++ packages/x-charts-premium/src/ChartsWebGLLayer/CanvasPositioner.tsx
@@ -1,0 +1,32 @@
+'use client';
+import * as React from 'react';
+import { selectorChartSvgHeight, selectorChartSvgWidth, useStore } from '@mui/x-charts/internals';
+
+export function CanvasPositioner({
+  children,
+  ...other
+}: React.PropsWithChildren<React.ComponentProps<'div'>>) {
+  const store = useStore();
+  const svgWidth = store.use(selectorChartSvgWidth);
+  const svgHeight = store.use(selectorChartSvgHeight);
+
+  return (
+    <div
+      {...other}
+      style={{
+        position: 'absolute',
+        inset: 0,
+        pointerEvents: 'none',
+        /* Ensures the canvas occupies the same space as the SVG */
+        maxWidth: svgWidth,
+        maxHeight: svgHeight,
+        width: '100%',
+        height: '100%',
+        margin: 'auto',
+      }}
+      aria-hidden
+    >
+      {children}
+    </div>
+  );
+}
